+++ DotNet/Util.Test/EquatableReferenceType.cs
