+++ DotNet/Util.Test/Extensions/EnumerableTests.cs
