+++ DotNet/Util.Test/Extensions/MaybeFirstTest.cs
