+++ DotNet/Util.Test/Extensions/ReferenceTypeNullableExtensionTestContext.cs
@@ -11,6 +11,7 @@
 // WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
 
+using DotNet.Util.Extensions;
 using System;
 using Xunit;
