+++ DotNet/Util.Test/Results/CommandResultTest.cs
