+++ DotNet/Util.Test/Results/QueryResultTest.cs
