+++ Util.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.29728.190
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Util", "DotNet\Util\Util.csproj", "{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Util.Test", "DotNet\Util.Test\Util.Test.csproj", "{CAB77004-D1D3-42FA-92C3-EFB93B13928F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1D6DC1E6-56D1-4A1D-8E91-227F1D85DA3B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CAB77004-D1D3-42FA-92C3-EFB93B13928F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
